New: FixedColumns now uses the column width defined by DataTables exclusively
This means that the options:
  - sLeftWidth
  - iLeftWidth
  - sRightWidth
  - iRightWidth
have all been removed.

No errors will be thrown when they are passed in, they simply will no
longer be used, with FixedColumns basing all of its column width
calculations on what DataTables has rendered before it.

This means that if you want a fixed column width, or a specific relative
one, apply it to your table using the width attribute for the column
header or the `width` option in DataTables.

Examples updated to take account of this.