Fix: Memory leak when redrawing the table as event listeners were not removed
1. Two cases of using direct DOM manipulation (`removeChild`) which have
jQuery events assigned to them (cloned from the DataTables' header for
the click to sort cells). Using `removeChild` wouldn't remove the jQuery
listener, so use `$().remove()` which is also shorter.

2. When cloning the headers to copy the class names, but not using the
elements, the cloned events would be left hanging. Added a flag to
`_fnCopyLayout` which will signal if events should be copied or not.