fix scroll bar if hide for first fixed column
FixedColumn is a powerful plugin to create nice grids. It greatly helps
to my work. Recent I found one issue when using fixed column:

If it hide the first left column(which is also included as fixed
column), there would be a x-scrollbar displayed for the left side.
Tested on my mac with mouse connected. If I use touch pad, it works
fine.

This is because border(usually 1px) missing in the calculation.
This is my first pull request. Please review.